+++ apworld/brotato/options.py
@@ -32,6 +32,21 @@ class NumberRequiredWins(Range):
     display_name = "Wins Required"
 
 
+class NumberCharacters(Range):
+    """The total number of characters in the world.
+
+    The characters will be randomly selected from the valid characters determined by the "Include Characters" options.
+
+    The actual number of characters may be less if there are not enough included characters in the options.
+    """
+
+    range_start = 1
+    range_end = TOTAL_NUM_CHARACTERS
+
+    default = 15
+    display_name = "Number of Included Characters"
+
+
 class StartingCharacters(Choice):
     """Determines your set of starting characters.
 
@@ -421,9 +436,10 @@ class IncludeAbyssalTerrorsCharacters(OptionSet):
 @dataclass
 class BrotatoOptions(PerGameCommonOptions):
     num_victories: NumberRequiredWins
+    num_characters: NumberCharacters
+    num_starting_characters: NumberStartingCharacters
     starting_characters: StartingCharacters
     include_base_game_characters: IncludeBaseGameCharacters
-    num_starting_characters: NumberStartingCharacters
     waves_per_drop: WavesPerCheck
     gold_reward_mode: GoldRewardMode
     xp_reward_mode: XpRewardMode