+++ apworld/brotato/test/data_sets/base.py
@@ -1,0 +1,13 @@
+from abc import ABC, abstractmethod
+from typing import Any
+
+
+class BrotatoTestDataSet(ABC):
+    @abstractmethod
+    def test_name(self) -> str:
+        pass
+
+    @property
+    @abstractmethod
+    def options_dict(self) -> dict[str, Any]:
+        pass