+++ apworld/brotato/test/test_num_victories.py
@@ -12,22 +12,21 @@ class TestBrotatoNumVictoriesOption(BrotatoTestBase):
 
         We also check the output slot data to make sure the change, if made, propagates to it as well.
         """
-        expected_final_num_victories_value = BASE_GAME_CHARACTERS.num_characters - 5
-        include_characters = BASE_GAME_CHARACTERS.characters[
-            :expected_final_num_victories_value
-        ]
-        original_num_victories_value = BASE_GAME_CHARACTERS.num_characters
+        expected_final_num_victories_value = 25  # Arbitrary value that's less than the number of characters
+        include_characters = BASE_GAME_CHARACTERS.characters
+        num_victories_option_value = 35
 
         options = {
-            "num_victories": original_num_victories_value,
+            "num_victories": num_victories_option_value,
             "include_base_game_characters": include_characters,
+            "num_characters": expected_final_num_victories_value,
         }
 
         self._run(options)
 
-        final_num_victories_value = self.world.options.num_victories.value
+        generated_num_victories_value = self.world.options.num_victories.value
         # Checking slot data here so we don't need to duplicate the test setup in multiple functions/files.
         slot_data_num_victories = self.world.fill_slot_data()["num_wins_needed"]
 
-        self.assertEqual(expected_final_num_victories_value, final_num_victories_value)
+        self.assertEqual(expected_final_num_victories_value, generated_num_victories_value)
         self.assertEqual(expected_final_num_victories_value, slot_data_num_victories)